function create_uncons_regionD updated
the function uncons_regionD returns a matrix with two columns mu3, mu4.

the function accepts vectors of mu3_p and mu4_p.
This is useful when generating heatmaps to explore the search space.

This is also good for gettting an structured return with named columns.

The csm_fit_* functions also return named vectors.